+++ app/src/main/java/org/nsh07/simplygraph/NativeBridge.kt
@@ -6,7 +6,7 @@ class NativeBridge {
     }
 
     external fun calculateGraphPoints(
-        xWidth: Int,
+        xWidth: Double,
         yWidth: Double,
         xOffset: Double,
         yOffset: Double,
