+++ app/src/main/java/org/nsh07/simplygraph/ui/viewModel/UiState.kt
@@ -9,7 +9,7 @@ data class GraphState(
     val invalidations: Int = 0,
     val canvasSize: Size = Size(0f, 0f),
     val points: List<Offset> = emptyList(),
-    val xWidth: Int = 10,
+    val xWidth: Float = 10f,
     val xOffset: Float = 0f,
     val yOffset: Float = 0f,
     val connectPoints: Boolean = true
